# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Start Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/public/api",
            "description": "main server"
        },
        {
            "url": "/laravel/backend/public/api",
            "description": "local server"
        }
    ],
    "paths": {
        "/student/register": {
            "post": {
                "tags": [
                    "Auth Student"
                ],
                "summary": "Register as a new Student",
                "operationId": "authRegister",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Registration model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "RegisterResult response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResultRegisterResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/student/login": {
            "post": {
                "tags": [
                    "Auth Student"
                ],
                "summary": "Login as a Student",
                "operationId": "authLogin",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Login model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Profile response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResultApi"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/teacher/register": {
            "post": {
                "tags": [
                    "Auth Teacher"
                ],
                "summary": "Register as a new Teacher",
                "operationId": "authRegister",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Registration model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "RegisterResult response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResultRegisterResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/teacher/login": {
            "post": {
                "tags": [
                    "Auth Teacher"
                ],
                "summary": "Login as a Teacher",
                "operationId": "authLogin",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Login model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginAdminPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Profile response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginAdminResultApi"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/teacher/profile": {
            "get": {
                "tags": [
                    "Auth Teacher"
                ],
                "summary": "Get profile details",
                "operationId": "d7dc5a8b9aeb35aff3f8403c6beaae0f",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ApiResultProfileResult response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResultProfileAdminResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/teacher/update-profile": {
            "post": {
                "tags": [
                    "Auth Teacher"
                ],
                "summary": "Edit profile content",
                "operationId": "4f885d40920c7265438873617df3c36c",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#components/schemas/EditProfileAdminRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ApiResultProfileResult response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/teacher/logout": {
            "post": {
                "tags": [
                    "Auth Teacher"
                ],
                "summary": "Log out from system",
                "operationId": "63f42e5108d3dfb28ab672da0f525fe2",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResultRegisterResult": {
                "title": "ApiResultRegisterResult model",
                "description": "ApiResultRegisterResult model",
                "properties": {
                    "result": {
                        "$ref": "#/components/schemas/RegisterResult"
                    },
                    "isOk": {
                        "title": "isOk",
                        "description": "Indicates if the response is ok or not",
                        "type": "boolean"
                    },
                    "message": {
                        "$ref": "#/components/schemas/ApiMessage"
                    }
                },
                "type": "object"
            },
            "LoginPayload": {
                "title": "LoginPayload model",
                "description": "LoginPayload model",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginResult": {
                "title": "LoginResult model",
                "description": "LoginResult model",
                "properties": {
                    "resultCode": {
                        "title": "resultCode",
                        "description": ">-\r\n    Code of resonse (response type)\r\n        1 => Login Accepted\r\n        2 => Invalid Password\r\n        3 => Account not Confirmed\r\n        4 => User is BANNED\r\n        5 => Account not Active\r\n        6 => User not found",
                        "type": "integer"
                    },
                    "resultText": {
                        "title": "resultText",
                        "description": "Result text to show for user",
                        "type": "string"
                    },
                    "token": {
                        "title": "token",
                        "description": "Access token",
                        "type": "string"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/ProfileResult"
                    }
                },
                "type": "object"
            },
            "LoginResultApi": {
                "title": "LoginResultApi model",
                "description": "LoginResultApi model",
                "properties": {
                    "result": {
                        "$ref": "#/components/schemas/LoginResult"
                    },
                    "isOk": {
                        "title": "isOk",
                        "description": "Indicates if the response is ok or not",
                        "type": "boolean"
                    },
                    "message": {
                        "$ref": "#/components/schemas/ApiMessage"
                    }
                },
                "type": "object"
            },
            "ProfileResult": {
                "title": "ProfileResult model",
                "description": "ProfileResult model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID",
                        "type": "integer"
                    },
                    "name": {
                        "title": "last_name",
                        "description": "Last Name",
                        "type": "string"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Phone",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string"
                    },
                    "region": {
                        "title": "region",
                        "description": "region",
                        "type": "string"
                    },
                    "city": {
                        "title": "city",
                        "description": "city",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string"
                    },
                    "image": {
                        "title": "image",
                        "description": "Image Url",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RegisterModel": {
                "title": "RegisterModel model",
                "description": "RegisterModel model",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string"
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RegisterResult": {
                "title": "RegisterResult model",
                "description": "RegisterResult model",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string"
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string"
                    },
                    "message": {
                        "title": "message",
                        "description": "Message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ApiResultProfileAdminResult": {
                "title": "ApiResultProfileAdminResult model",
                "description": "ApiResultProfileAdminResult model",
                "properties": {
                    "result": {
                        "$ref": "#/components/schemas/ProfileAdminResult"
                    },
                    "isOk": {
                        "title": "isOk",
                        "description": "Indicates if the response is ok or not",
                        "type": "boolean"
                    },
                    "message": {
                        "$ref": "#/components/schemas/ApiMessage"
                    }
                },
                "type": "object"
            },
...
            "LoginAdminPayload": {
                "title": "LoginAdminPayload model",
                "description": "LoginAdminPayload model",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "Account Password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginAdminResult": {
                "title": "LoginAdminResult model",
                "description": "LoginAdminResult model",
                "properties": {
                    "resultCode": {
                        "title": "resultCode",
                        "description": ">-\r\n    Code of resonse (response type)\r\n        1 => Login Accepted\r\n        2 => Invalid Password\r\n        3 => Account not Confirmed\r\n        4 => User is BANNED\r\n        5 => Account not Active\r\n        6 => User not found",
                        "type": "integer"
                    },
                    "resultText": {
                        "title": "resultText",
                        "description": "Result text to show for user",
                        "type": "string"
                    },
                    "token": {
                        "title": "token",
                        "description": "Access token",
                        "type": "string"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/ProfileAdminResult"
                    }
                },
                "type": "object"
            },
            "LoginAdminResultApi": {
                "title": "LoginAdminResultApi model",
                "description": "LoginAdminResultApi model",
                "properties": {
                    "result": {
                        "$ref": "#/components/schemas/LoginAdminResult"
                    },
                    "isOk": {
                        "title": "isOk",
                        "description": "Indicates if the response is ok or not",
                        "type": "boolean"
                    },
                    "message": {
                        "$ref": "#/components/schemas/ApiMessage"
                    }
                },
                "type": "object"
            },
            "ProfileAdminResult": {
                "title": "ProfileResult model",
                "description": "ProfileResult model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID",
                        "type": "integer"
                    },
                    "name": {
                        "title": "last_name",
                        "description": "Last Name",
                        "type": "string"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Phone",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string"
                    },
                    "region": {
                        "title": "region",
                        "description": "region",
                        "type": "string"
                    },
                    "city": {
                        "title": "city",
                        "description": "city",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string"
                    },
                    "image": {
                        "title": "image",
                        "description": "Image Url",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ApiMessage": {
                "title": "ApiMessage model",
                "description": "ApiMessage model",
                "properties": {
                    "type": {
                        "title": "type",
                        "description": "Message type",
                        "type": "string"
                    },
                    "code": {
                        "title": "code",
                        "description": "Message code",
                        "type": "number"
                    },
                    "content": {
                        "title": "content",
                        "description": "Message content",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ApiResult": {
                "title": "ApiResult model",
                "description": "ApiResult model",
                "properties": {
                    "isOk": {
                        "title": "isOk",
                        "description": "Indicates if the response is ok or not",
                        "type": "boolean"
                    },
                    "message": {
                        "$ref": "#/components/schemas/ApiMessage"
                    },
                    "exception": {
                        "title": "exception",
                        "description": "Exception Message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Authentication Bearer Token",
                "name": "Authentication Bearer Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}